Refactor navigation structure and update dependencies
