+++ sample/composeApp/src/commonMain/kotlin/composegears/tiamat/sample/content/advanced/AdvBackStackAlteration.kt
@@ -1,18 +1,16 @@
 package composegears.tiamat.sample.content.advanced
 
-import androidx.compose.foundation.border
-import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.KeyboardArrowLeft
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import com.composegears.tiamat.compose.*
 import composegears.tiamat.sample.ui.*
@@ -20,7 +18,10 @@ import org.jetbrains.compose.ui.tooling.preview.Preview
 
 val AdvBackStackAlteration by navDestination(ScreenInfo()) {
     Screen("Back stack alteration") {
-        Column(
+        // todo migrate into edit-nav-state
+        //  add an example of edit current entry
+        //  with a diff nev types (back, forward, instant)
+        /*Column(
             modifier = Modifier.padding(16.dp),
             horizontalAlignment = Alignment.CenterHorizontally
         ) {
@@ -29,7 +30,7 @@ val AdvBackStackAlteration by navDestination(ScreenInfo()) {
                 startDestination = AdvBackStackAlterationScreenA,
             )
             val currentDestination by nc.currentNavDestinationAsState()
-            val backStack by nc.currentBackStackFlow.collectAsState()
+            val backStack by nc.navStackAsState()
             VSpacer()
             Text(
                 text = "Here some simple examples of backStack editing",
@@ -106,7 +107,7 @@ val AdvBackStackAlteration by navDestination(ScreenInfo()) {
                         shape = RoundedCornerShape(8.dp)
                     )
             )
-        }
+        }*/
     }
 }
